+++ src/components/ChefOrdersTable/ChefOrdersTable.jsx
@@ -5,15 +5,15 @@ import { GridRowEditStopReasons, GridRowModes } from '@mui/x-data-grid';
 import { chefsAmountAfterFee } from '@/helpers';
 import useChefOrder from '@/hooks/useChefOrders';
 import AppDataGridTable from '@/shared/AppDataGridTable/AppDataGridTable';
-import { formatDateForDataGrid } from '../UsersOrdersTable/formatDateForDataGrid';
-import { CustomPagination } from '../UsersOrdersTable/Pagination';
+import { formatDateForDataGrid } from '../UserOrdersTable/formatDateForDataGrid';
+import { CustomPagination } from '../UserOrdersTable/Pagination';
 import { getActions } from './getActions';
-import { getStatusOptions } from './getChefsStatusOptions';
+import { getStatusOptions } from './getChefStatusOptions';
 import { OrderItemsCell } from './OrderItemsCell';
 import { processRowUpdate } from './processRowUpdate';
 import { StatusCell } from './StatusCell';
 
-export const ChefsOrdersTable = () => {
+export const ChefOrdersTable = () => {
   const chefID = '6557219bccbbbbc3695bc8b2';
   const { data, isLoading, error } = useChefOrder(chefID);
