+++ src/components/UserOrdersTable/calculateTotalSum.js
